ui: improve modal sizing
