+++ src/libtrx/game/ui/elements/modal.c
@@ -15,7 +15,7 @@ static const UI_WIDGET_OPS m_Ops = {
 static void M_Measure(UI_NODE *const node)
 {
     node->measure_w = UI_GetCanvasWidth();
-    node->measure_h = UI_GetCanvasHeight() - TEXT_HEIGHT_FIXED;
+    node->measure_h = UI_GetCanvasHeight();
 }
 
 void UI_BeginModal(const float x, const float y)
