+++ src/tr1/meson.build
@@ -260,7 +260,6 @@ sources = [
   'game/spawn.c',
   'game/stats/common.c',
   'game/text.c',
-  'game/ui/common.c',
   'game/ui/dialogs/stats.c',
   'game/viewport.c',
   'global/enum_map.c',
